fedora/pyproject.toml
[tool.poetry]
name = "fedora"
version = "0.1.0"
description = ""
authors = ["Jim Crews <crews.jim@gmail.com>"]
readme = "README.md"
packages = [{include = "*", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.2.5"
djangorestframework = "^3.16.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
